refactor: OrderTable - TableGroup 간 객체참조 제거
