+++ src/main/java/kitchenpos/domain/tableGroup/TableGroup.java
@@ -1,15 +1,11 @@
-package kitchenpos.domain;
+package kitchenpos.domain.tableGroup;
 
-import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 public class TableGroup {
@@ -19,18 +15,13 @@ public class TableGroup {
     private Long id;
     @Column(name = "created_date")
     private LocalDateTime createdDate;
-    @OneToMany(cascade = CascadeType.PERSIST,mappedBy = "tableGroup")
-    private List<OrderTable> orderTables;
 
     public TableGroup() {
     }
 
-    public TableGroup(LocalDateTime createdDate, List<OrderTable> orderTables) {
+
+    public TableGroup(LocalDateTime createdDate) {
         this.createdDate = createdDate;
-        for (OrderTable orderTable : orderTables) {
-            orderTable.addTableGroup(this);
-        }
-        this.orderTables = orderTables;
     }
 
     public Long getId() {
@@ -43,13 +34,5 @@ public class TableGroup {
 
     public LocalDateTime getCreatedDate() {
         return createdDate;
-    }
-
-    public List<OrderTable> getOrderTables() {
-        return orderTables;
-    }
-
-    public void addOrderTable(OrderTable orderTable){
-        orderTables.add(orderTable);
     }
 }
